requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
zhipuai
PyPDF2
faiss-cpu
python-dotenv